refactor(package): package 정리
