+++ src/main/java/lotto/constant/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.common;
+package lotto.constant;
 
 public class Lotto {
     private Lotto() {
