+++ src/main/java/lotto/utils/StringUtils.java
@@ -1,6 +1,6 @@
 package lotto.utils;
 
-import lotto.common.ErrorMessage;
+import lotto.constant.ErrorMessage;
 
 import java.util.Arrays;
 import java.util.List;
