+++ src/test/java/lotto/IntegerUtilsTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
 
-import lotto.common.ErrorMessage;
+import lotto.constant.ErrorMessage;
 import lotto.utils.IntegerUtils;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
