+++ src/test/java/lotto/StringUtilsTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.common.ErrorMessage;
+import lotto.constant.ErrorMessage;
 import lotto.utils.StringUtils;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
